Created JoinGame method in ClientMethods and fixed Login and CreateGame methods.
Created joinParser to return an object of the game class for JoinGame method to read.

Made SnakeMasterId variable in game class to fit new server requirements. The server returned a winner as boolean which didn't work out. Now it returns a boolean AND a SnakeMasterId which is given to the winner. Now it actually is possible to win (Still needs a little fix though..).